Hoist user fetching out of child views
In order to be able to share user state between the standard and expanded student views while also being able to clear that state when navigating away from either page, these changes add a new parent component for the both of them which handles fetching user information and dispatching it to state. This way, a cleanup function to clear that state can instead be placed within the parent function allowing both views to share state without having to re-request each time.