Apply useForm to AddPlanPopper
As a proof-of-concept, I converted the AddPlanPopper to use the useForm hook just to demonstrate how it works, which made things a bit cleaner by being able to separate the validation state and logic out of the main component. There's a bit of boilerplate for setting initial values and setting up the validation logic, but this all seems necessary.