Remove init errors and allow undefined error
This is actually not really necessary and makes setting up the validator less intuitive. If we just allow errors to be undefined, then there's no point in having a set error type per form and we don't need to remember to return the empty string at the end of every validation function.