Proof-of-concept to add useForm hook for handling form state and validation AddPlanPopper (#338)
* Add useForm hook

This hook is meant to contain both form field state and error state. It abstracts out some simple core form validation functionalities that were used across the existing forms in the app such as validating fields, showing errors on submit, and resetting the form.

* Apply useForm to AddPlanPopper

As a proof-of-concept, I converted the AddPlanPopper to use the useForm hook just to demonstrate how it works, which made things a bit cleaner by being able to separate the validation state and logic out of the main component. There's a bit of boilerplate for setting initial values and setting up the validation logic, but this all seems necessary.

* Remove init errors and allow undefined error

This is actually not really necessary and makes setting up the validator less intuitive. If we just allow errors to be undefined, then there's no point in having a set error type per form and we don't need to remember to return the empty string at the end of every validation function.

* Get rid of multiple validators bc YAGNI