bugfix: make transfer courses add correctly to schedule
Fixes a bug where transfer courses added during onboarding weren't correctly
being added to the schedule as transfer courses.

This commit fixes a variety of bugs:

1. slowdown due to consecutive awaits for fetching courses (fix is to use Promise.all)
2. incorrect course equality checks
3. double-equals equality checks

Additionally, this cleans up some missing-key warnings, deduplicates some code,
and also adds text for the steps displayed during onboarding.

Outstanding work may be to change the strategy in which courses are added to the schedule.

Changes:

- refactor TransferCoursesScreen to be functional component
- pull out shared utils from transfer screen and completed courses screen to course-helpers
- add explicitly controlled SelectableCourse component to GenericOnboarding